+++ packages/babel-preset-sui/src/clean-list.js
@@ -1,3 +1,0 @@
-module.exports = function(list) {
-  return list.filter(Boolean)
-}