+++ packages/sui-mono/bin/sui-mono-release.js
@@ -2,7 +2,6 @@
 require('util.promisify/shim')()
 const program = require('commander')
 const path = require('path')
-const {shell} = require('@tunnckocore/execa')
 const config = require('../src/config')
 const checker = require('../src/check')
 const {serialSpawn, showError} = require('@s-ui/helpers/cli')
@@ -17,6 +16,7 @@ program
   .option('-T, --github-token <token>', 'github token')
   .option('-U, --github-user <user>', 'github user')
   .option('-E, --github-email <email>', 'github email')
+  .option('--skip-ci', 'Add [skip ci] to release commit message', false)
   .on('--help', () => {
     console.log('  Description:')
     console.log('')
@@ -67,7 +67,7 @@ const singlePackageRelease = ({status, packageScope}) =>
     .filter(scope => scope === packageScope)
     .map(scope => scopeMapper({scope, status}))
 
-const releaseEachPkg = ({pkg, code} = {}) => {
+const releaseEachPkg = ({pkg, code} = {}, {skipCI} = {}) => {
   return new Promise((resolve, reject) => {
     if (code === 0) {
       return resolve()
@@ -104,10 +104,15 @@ const releaseEachPkg = ({pkg, code} = {}) => {
       .then(() => {
         // Create release commit
         const {version} = getPackageJson(cwd, true)
-        return serialSpawn(
-          [['git', [`commit -m "release(${packageScope}): v${version}"`]]],
-          {cwd}
-        )
+
+        // We're adding [skip ci] to the commit message to avoid
+        // start a build on CI if not needed
+        // docs: https://docs.travis-ci.com/user/customizing-the-build/#skipping-a-build
+        const commitMsg = `release(${packageScope}): v${version}${
+          skipCI ? ' [skip ci]' : ''
+        }`
+
+        return serialSpawn([['git', [`commit -m "${commitMsg}"`]]], {cwd})
       })
       .then(() => serialSpawn(docCommands))
       .then(() => {
@@ -129,11 +134,11 @@ const releaseMode = packageScope ? singlePackageRelease : releasesByPackages
 
 const checkIsMasterBranchActive = async ({status, cwd}) => {
   try {
-    const output = await exec(`git rev-parse --abbrev-ref HEAD`, {
+    const {stdout} = await exec(`git rev-parse --abbrev-ref HEAD`, {
       cwd
     })
 
-    if (output.stdout.trim() === 'master') {
+    if (stdout.trim() === 'master') {
       return Promise.resolve(status)
     } else {
       throw new Error(
@@ -147,27 +152,14 @@ const checkIsMasterBranchActive = async ({status, cwd}) => {
   }
 }
 
-const execute = async (cmd, full) => {
-  try {
-    console.log('--->', cmd)
-    const [resp] = await shell(cmd)
-    const output = full ? resp : resp.stdout
-    console.log(output)
-    return output
-  } catch (e) {
-    const output = full ? e : e.stderr
-    console.log(output)
-    return e
-  }
-}
-
 const automaticRelease = async ({
   githubToken,
   githubUser,
   githubEmail,
   cwd
 }) => {
-  const repoURL = await execute('git config --get remote.origin.url')
+  const {stdout} = await exec('git config --get remote.origin.url', {cwd})
+  const repoURL = stdout.trim()
   const gitURL = gitUrlParse(repoURL).toString('https')
   const authURL = new URL(gitURL)
   authURL.username = githubToken
@@ -188,7 +180,7 @@ const isAutomaticRelease = ({githubToken, githubUser, githubEmail}) => {
 checker
   .check()
   .then(async status => {
-   const {githubEmail, githubToken, githubUser} = program
+    const {githubEmail, githubToken, githubUser} = program
     isAutomaticRelease({
       githubEmail,
       githubToken,
@@ -207,11 +199,15 @@ checker
   .then(releases =>
     releases
       .filter(({code}) => code !== 0)
-      .map(release => () => releaseEachPkg(release))
+      .map(release => () => releaseEachPkg(release, {skipCI: program.skipCi}))
       // https://gist.github.com/istarkov/a42b3bd1f2a9da393554
       .reduce(
         (m, p) => m.then(v => Promise.all([...v, p()])),
         Promise.resolve([])
       )
   )
-  .catch(console.log.bind(console))
+  .catch(err => {
+    console.error('[sui-mono release] ERROR:')
+    console.error(err)
+    process.exit(1)
+  })