+++ packages/sui-studio/src/components/demo/Tab.js
@@ -1,16 +1,15 @@
-import PropTypes from 'prop-types'
 import React from 'react'
+import PropTypes from 'prop-types'
 import cx from 'classnames'
 
 export default function Tab({handleClick, isActive, literal}) {
-  const getItemClassName = ({isActive}) =>
-    cx('sui-StudioTabs-button', {
-      'sui-StudioTabs-button--active': isActive
-    })
+  const className = cx('sui-StudioTabs-button', {
+    'sui-StudioTabs-button--active': isActive
+  })
 
   return (
     <li className="sui-StudioTabs-tab">
-      <button className={getItemClassName({isActive})} onClick={handleClick}>
+      <button className={className} onClick={handleClick}>
         {literal}
       </button>
     </li>