+++ packages/sui-studio/src/components/demo/ThemesButtons.js
@@ -1,5 +1,5 @@
-import PropTypes from 'prop-types'
 import React from 'react'
+import PropTypes from 'prop-types'
 import Tabs from './Tabs'
 import Tab from './Tab'
 
@@ -10,7 +10,7 @@ const ThemesButtons = ({themes, onThemeChange, selected}) => {
     <Tabs title="Theme">
       {['default', ...themes].map((theme, index) => (
         <Tab
-          handleClick={evt => onThemeChange(theme, index)}
+          handleClick={() => onThemeChange(theme, index)}
           key={`${theme}${index}`}
           isActive={index === selected}
           literal={theme}
@@ -20,7 +20,6 @@ const ThemesButtons = ({themes, onThemeChange, selected}) => {
   )
 }
 
-ThemesButtons.displayName = 'ThemesButtons'
 ThemesButtons.propTypes = {
   themes: PropTypes.array,
   onThemeChange: PropTypes.func,