+++ packages/sui-studio/src/components/documentation/Api.js
@@ -1,34 +1,35 @@
 import PropTypes from 'prop-types'
 import React, {useEffect, useState} from 'react'
-import {tryRequireRawSrc} from '../tryRequire'
+import {fetchComponentSrcRawFile} from '../tryRequire'
 
 export default function Api({params}) {
-  const [docs, setDocs] = useState(false)
+  const [docs, setDocs] = useState(null)
+
   useEffect(() => {
     async function getDocs() {
       const reactDocs = await import('react-docgen')
-      const src = await tryRequireRawSrc(params)
-      const docs = reactDocs.parse(src)
+      const {category, component} = params
+      const rawSource = await fetchComponentSrcRawFile({category, component})
+      const docs = reactDocs.parse(rawSource)
       setDocs(docs)
     }
     getDocs()
-  })
+  }, [params])
 
-  const renderPropsApi = ({propsApi = {}}) => {
-    const keysOfProps = Object.keys(propsApi).sort((a, b) => a.localeCompare(b))
+  const renderPropsApi = ({props = {}}) => {
+    const keysOfProps = Object.keys(props).sort((a, b) => a.localeCompare(b))
     // if the component doesn't have props, show a message
     if (keysOfProps.length === 0) {
       return <p>This component doesn't have props</p>
     }
     // if we have props, render all of them using React
-    const renderedProps = keysOfProps.map(propName => {
-      const {defaultValue = {}, required, type, description} = propsApi[
-        propName
-      ]
+    return keysOfProps.map(propName => {
+      const {defaultValue = {}, required, type, description} = props[propName]
       const {value = undefined} = defaultValue
 
       if (typeof type === 'undefined') {
-        console.warn( // eslint-disable-line
+        console.warn(
+          // eslint-disable-line
           'It seem that you might have a prop with a defaultValue but it does not exist as propType'
         )
         return
@@ -37,7 +38,7 @@ export default function Api({params}) {
       return (
         <div className="sui-StudioProps-prop" key={propName}>
           <h3>{propName}</h3>
-          <div className="sui-StudioProps-tags">
+          <div>
             <div className="sui-StudioProps-tag sui-StudioProps-required">
               <span>required</span>
               <span className={required ? 'is-required' : ''}>
@@ -59,19 +60,21 @@ export default function Api({params}) {
         </div>
       )
     })
-    // return all the rendered props with a title
-    return [<h2 key="propTitles">Props</h2>, ...renderedProps]
   }
 
   if (docs) {
     const {props} = docs
-    return renderPropsApi({propsApi: props})
+    return (
+      <>
+        <h2>Props</h2>
+        {renderPropsApi({props})}
+      </>
+    )
   }
 
   return null
 }
 
-Api.displayName = 'Api'
 Api.propTypes = {
   params: PropTypes.shape({
     category: PropTypes.string,