+++ packages/sui-studio/workbench/src/components/Raw/index.js
@@ -4,23 +4,19 @@ import PropTypes from 'prop-types'
 import SUIContext from '@s-ui/react-context'
 
 import Preview from '../../../../src/components/preview'
-import withContext from '../../../../src/components/demo/HoC/withContext'
 import Style from '../../../../src/components/style'
-import When from '../../../../src/components/when'
 import {
   createContextByType,
   cleanDisplayName,
-  pipe,
   removeDefaultContext
 } from '../../../../src/components/demo/utilities'
 
 import Component, * as named from 'component'
 
 import './index.scss'
-
 let playground
 try {
-  playground = require('!raw-loader!demo/playground')
+  playground = require('!raw-loader!demo/playground').default
 } catch (e) {}
 
 const nonDefault = removeDefaultContext(named)
@@ -38,10 +34,6 @@ export default function Raw({
 
   // check if is a normal component or it's wrapped with a React.memo method
   const ComponentToRender = Component.type ? Component.type : Component
-  const Enhance = pipe(withContext(context, context))(ComponentToRender)
-
-  const EnhanceDemoComponent =
-    DemoComponent && pipe(withContext(context, context))(DemoComponent)
 
   return (
     <div className="Raw">
@@ -49,22 +41,22 @@ export default function Raw({
       <Style id="sui-studio-raw-theme">{themes[actualStyle]}</Style>
 
       <div className="Raw-center">
-        <When value={!EnhanceDemoComponent && playground}>
-          {() => (
-            <Preview
-              scope={{
-                context,
-                React,
-                [cleanDisplayName(Enhance.displayName)]: Enhance,
-                ...nonDefault
-              }}
-              code={playground}
-            />
-          )}
-        </When>
-        {EnhanceDemoComponent && (
+        {!DemoComponent && playground && (
+          <Preview
+            scope={{
+              context,
+              React,
+              [cleanDisplayName(
+                ComponentToRender.displayName
+              )]: ComponentToRender,
+              ...nonDefault
+            }}
+            code={playground}
+          />
+        )}
+        {DemoComponent && (
           <SUIContext.Provider value={context}>
-            <EnhanceDemoComponent />
+            <DemoComponent />
           </SUIContext.Provider>
         )}
       </div>