+++ packages/sui-studio/workbench/src/components/Root/index.js
@@ -1,26 +1,9 @@
-/* global __EXPERIMENTAL_TEST_DEV__ */
 import React, {useState} from 'react'
 import PropTypes from 'prop-types'
 
 import Header from '../Header'
 import Select from '../Select'
-import Test from '../../../../src/components/test'
-import When from '../../../../src/components/when'
-
-const DEVICES = {
-  mobile: {
-    width: 362,
-    height: 642
-  },
-  tablet: {
-    width: 768,
-    height: 1024
-  },
-  desktop: {
-    width: '100%',
-    height: '100%'
-  }
-}
+import Test from '../Test'
 
 const importComponent = () => import('component/index.js')
 const importTest = () => import('test/index.js')
@@ -34,69 +17,47 @@ const updateOnChange = (setState, sessionKey) => nextValue => {
 }
 
 export default function Root({componentID, contexts = {}, themes}) {
-  const [actualContext, setActualContext] = useState(
+  const [actualContext, setActualContext] = useState(() =>
     getFromStorage('actualContext', 'default')
   )
-  const [actualStyle, setActualStyle] = useState(
+  const [actualStyle, setActualStyle] = useState(() =>
     getFromStorage('actualStyle', 'default')
   )
-  const [actualDevice, setActualDevice] = useState(
-    getFromStorage('actualDevice', 'mobile')
-  )
-
-  const [showTests, setShowTests] = useState(
+  const [showTests, setShowTests] = useState(() =>
     getFromStorage('showTests', 'show')
   )
 
-  const {width, height} = DEVICES[actualDevice]
   const iframeSrc = `/?raw=true&actualStyle=${actualStyle}&actualContext=${actualContext}`
 
   return (
     <div className="Root">
-      <div className="Root-top">
-        <Header componentID={componentID} iframeSrc={iframeSrc}>
-          <Select
-            label="Contexts"
-            options={contexts}
-            initValue={actualContext}
-            onChange={updateOnChange(setActualContext, 'actualContext')}
-          />
-          <Select
-            label="Themes"
-            options={themes}
-            initValue={actualStyle}
-            onChange={updateOnChange(setActualStyle, 'actualStyle')}
-          />
-          <Select
-            label="Devices"
-            options={DEVICES}
-            initValue={actualDevice}
-            onChange={updateOnChange(setActualDevice, 'actualDevice')}
-          />
-        </Header>
-      </div>
-      <div className={`Root-center Root-${actualDevice}`}>
-        <span className={`Root-${actualDevice}-camera`} />
-        <span className={`Root-${actualDevice}-speaker`} />
-        <span className={`Root-${actualDevice}-button`} />
-        <iframe
-          className="Root-iframe"
-          style={{width, height}}
-          src={iframeSrc}
-          scrolling="yes"
+      <Header componentID={componentID} iframeSrc={iframeSrc}>
+        <Select
+          label="contexts"
+          options={contexts}
+          initValue={actualContext}
+          onChange={updateOnChange(setActualContext, 'actualContext')}
         />
-      </div>
-      <button
-        className={`Root-testSwitch--${showTests}`}
-        onClick={() => {
-          updateOnChange(
-            setShowTests,
-            'showTests'
-          )(showTests === 'show' ? 'hide' : 'show')
-        }}
-      >
-        {showTests === 'show' ? 'Close Tests' : 'Open Tests'}
-      </button>
+        <Select
+          label="themes"
+          options={themes}
+          initValue={actualStyle}
+          onChange={updateOnChange(setActualStyle, 'actualStyle')}
+        />
+        <button
+          className="Root-testSwitch"
+          onClick={() => {
+            updateOnChange(
+              setShowTests,
+              'showTests'
+            )(showTests === 'show' ? 'hide' : 'show')
+          }}
+        >
+          {showTests === 'show' ? 'Close Tests' : 'Open Tests'}
+        </button>
+      </Header>
+
+      <iframe src={iframeSrc} scrolling="yes" title="Demo" />
       <div className="Root-test" hidden={showTests === 'hide'}>
         <Test
           open