+++ packages/babel-preset-sui/CHANGELOG.md
@@ -2,6 +2,33 @@
 
 All notable changes to this project will be documented in this file.
 
+# 3.12.0 (2020-09-28)
+
+
+### Features
+
+* add babel-plugin-preval plugin to unlock some further optimizations to our c ([c7dcd5a](https://github.com/SUI-Components/sui/commit/c7dcd5a0f47bd7fe83f424d1bd37996e0dff3e07))
+
+
+
+# 3.11.0 (2020-09-17)
+
+
+### Features
+
+* simplify code and remove not needed files ([3d71d68](https://github.com/SUI-Components/sui/commit/3d71d682a4ff0668d9e99768aa50a0b0c4d31749))
+
+
+
+# 3.10.0 (2020-08-31)
+
+
+### Features
+
+* upgrade babel dependencies ([32ec90d](https://github.com/SUI-Components/sui/commit/32ec90d18cbb34b035b0bd6092af7d68a89965f3))
+
+
+
 # 3.9.0 (2020-07-06)