+++ packages/sui-svg/CHANGELOG.md
@@ -2,6 +2,24 @@
 
 All notable changes to this project will be documented in this file.
 
+# 2.6.0 (2020-09-28)
+
+
+### Features
+
+* stop using deprecated no-pre-loader param ([93ce122](https://github.com/SUI-Components/sui/commit/93ce1227fc2956ea9862095dccae4e47bb480e28))
+
+
+
+# 2.5.0 (2020-09-16)
+
+
+### Features
+
+* upgrade to latest sui-bundler version ([f3a4577](https://github.com/SUI-Components/sui/commit/f3a4577bf190f2e5c1d290e0a3b81e607bc49bd1))
+
+
+
 # 2.4.0 (2020-06-10)